Split logical planning into caps-logical module
