+++ caps-cora/src/test/scala/org/opencypher/caps/impl/logical/LogicalPlannerTest.scala
@@ -22,7 +22,7 @@ import org.opencypher.caps.api.schema.Schema
 import org.opencypher.caps.api.types._
 import org.opencypher.caps.api.value.{CypherBoolean, CypherInteger, CypherString}
 import org.opencypher.caps.impl.logical
-import org.opencypher.caps.impl.util.VarConverters.toVar
+import org.opencypher.caps.ir.impl.util.VarConverters.toVar
 import org.opencypher.caps.ir.api._
 import org.opencypher.caps.ir.api.block._
 import org.opencypher.caps.ir.api.expr._
