+++ caps-logical/src/main/scala/org/opencypher/caps/impl/logical/LogicalOptimizer.scala
